attention testing

+++ inc/kernals/attention.h
@@ -3,6 +3,6 @@
 #include<stdio.h>
 #include<stdint.h>
 #include<stdbool.h>
-void attention(double* Q, double* K, double* V, uint16_t n, uint16_t d, double* out, 
+void attention(double* Q, double* K, double* V, uint32_t n, uint32_t d, double* out, 
                  bool masked);
 #endif
